ReadLEULong problem with Zip64 and typecasting fixed. (as reported in http://community.icsharpcode.net/forums/thread/16543.aspx)

+++ src/Zip/ZipFile.cs
@@ -2753,7 +2753,7 @@ namespace ICSharpCode.SharpZipLib.Zip
 
 		ulong ReadLEUlong()
 		{
-			return ReadLEUint() | (ReadLEUint() << 32);
+			return ReadLEUint() | ((ulong)ReadLEUint() << 32);
 		}
 
 		#endregion
